+++ ts/src/lib/push-values/constant.ts
@@ -3,16 +3,16 @@ import { CaseStyle } from "kryo/case-style";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { LiteralType } from "kryo/types/literal";
 import { Uint16 } from "semantic-types";
-import { $ValueType, ValueType } from "../value-type";
+import { $PushValueType, PushValueType } from "../push-value-type";
 
 export interface Constant {
-  type: ValueType.Constant;
+  type: PushValueType.Constant;
   value: Uint16;
 }
 
 export const $Constant: DocumentIoType<Constant> = new DocumentType<Constant>({
   properties: {
-    type: {type: new LiteralType({type: $ValueType, value: ValueType.Constant as ValueType.Constant})},
+    type: {type: new LiteralType({type: $PushValueType, value: PushValueType.Constant as PushValueType.Constant})},
     value: {type: $Uint16},
   },
   changeCase: CaseStyle.SnakeCase,