+++ ts/src/lib/push-values/null.ts
@@ -1,15 +1,15 @@
 import { CaseStyle } from "kryo/case-style";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { LiteralType } from "kryo/types/literal";
-import { $ValueType, ValueType } from "../value-type";
+import { $PushValueType, PushValueType } from "../push-value-type";
 
 export interface Null {
-  type: ValueType.Null;
+  type: PushValueType.Null;
 }
 
 export const $Null: DocumentIoType<Null> = new DocumentType<Null>({
   properties: {
-    type: {type: new LiteralType({type: $ValueType, value: ValueType.Null as ValueType.Null})},
+    type: {type: new LiteralType({type: $PushValueType, value: PushValueType.Null as PushValueType.Null})},
   },
   changeCase: CaseStyle.SnakeCase,
 });