+++ ts/src/lib/push-values/sint32.ts
@@ -3,16 +3,16 @@ import { CaseStyle } from "kryo/case-style";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { LiteralType } from "kryo/types/literal";
 import { Sint32 as _Sint32 } from "semantic-types";
-import { $ValueType, ValueType } from "../value-type";
+import { $PushValueType, PushValueType } from "../push-value-type";
 
 export interface Sint32 {
-  type: ValueType.Sint32;
+  type: PushValueType.Sint32;
   value: _Sint32;
 }
 
 export const $Sint32: DocumentIoType<Sint32> = new DocumentType<Sint32>(() => ({
   properties: {
-    type: {type: new LiteralType({type: $ValueType, value: ValueType.Sint32 as ValueType.Sint32})},
+    type: {type: new LiteralType({type: $PushValueType, value: PushValueType.Sint32 as PushValueType.Sint32})},
     value: {type: _$Sint32},
   },
   changeCase: CaseStyle.SnakeCase,