+++ ts/src/lib/push-values/undefined.ts
@@ -1,15 +1,15 @@
 import { CaseStyle } from "kryo/case-style";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { LiteralType } from "kryo/types/literal";
-import { $ValueType, ValueType } from "../value-type";
+import { $PushValueType, PushValueType } from "../push-value-type";
 
 export interface Undefined {
-  type: ValueType.Undefined;
+  type: PushValueType.Undefined;
 }
 
 export const $Undefined: DocumentIoType<Undefined> = new DocumentType<Undefined>({
   properties: {
-    type: {type: new LiteralType({type: $ValueType, value: ValueType.Undefined as ValueType.Undefined})},
+    type: {type: new LiteralType({type: $PushValueType, value: PushValueType.Undefined as PushValueType.Undefined})},
   },
   changeCase: CaseStyle.SnakeCase,
 });